Add new conftest.py and asserts for metrics
