+++ models/experimental/stable_diffusion_xl_base/tests/conftest.py
@@ -1,0 +1,37 @@
+# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+
+# SPDX-License-Identifier: Apache-2.0
+
+import pytest
+
+
+def pytest_addoption(parser):
+    parser.addoption(
+        "--start-from",
+        action="store",
+        default=0,
+        help="Start from prompt number (0-4999)",
+    )
+    parser.addoption(
+        "--num-prompts",
+        action="store",
+        default=5000,
+        help="Number of prompts to process (default: 5000)",
+    )
+
+
+@pytest.fixture
+def accuracy_eval_range(request):
+    start_from = request.config.getoption("--start-from")
+    num_prompts = request.config.getoption("--num-prompts")
+    if start_from is not None:
+        start_from = int(start_from)
+    else:
+        start_from = 0
+
+    if num_prompts is not None:
+        num_prompts = int(num_prompts)
+    else:
+        num_prompts = 5000
+
+    return start_from, num_prompts
